# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APP API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/sms/verify": {
            "post": {
                "summary": "Post a 5 digit verification PIN, Headers MUST contain the authorization bearer token.",
                "operationId": "App\\Http\\Controllers\\API\\v1\\Authentication\\VerificationController::verifySMS",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "verification_code": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "verification_code": 24465
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "'verification': true"
                    }
                }
            }
        },
...
        "/api/v1/business-posts": {
            "put": {
                "summary": "Update business post by ID",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessPostsController::update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "id of business post",
                                        "type": "string"
                                    },
                                    "business_id": {
                                        "description": "id of business",
                                        "type": "string"
                                    },
                                    "text": {
                                        "description": "id of user",
                                        "type": "string"
                                    },
                                    "photo": {
                                        "description": "base64 encoded image",
                                        "type": "string"
                                    },
                                    "expire_date": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Business post updated"
                    },
                    "400": {
                        "description": "Invalid given data"
                    },
                    "404": {
                        "description": "Business post not found"
                    }
                }
            },
            "post": {
                "summary": "Create business post",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessPostsController::store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "business_id": {
                                        "description": "id of business",
                                        "type": "string"
                                    },
                                    "text": {
                                        "description": "id of user",
                                        "type": "string"
                                    },
                                    "photo": {
                                        "description": "base64 encoded image",
                                        "type": "string"
                                    },
                                    "expire_date": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "BusinessPostResource"
                    }
                }
            },
            "delete": {
                "summary": "Delete a business post",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessPostsController::destroy",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "id of business post",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Business Post Resource"
                    }
                }
            }
        },
        "/api/v1/business-posts/business/{$businessUuid}": {
            "get": {
                "summary": "Get all business posts for business",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessPostsController::forBusiness",
                "parameters": [
                    {
                        "name": "business_id",
                        "in": "query",
                        "description": "ID of business",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Business Post"
                    }
                }
            }
        },
...
        "/api/v1/business-reviews/{business}": {
            "post": {
                "summary": "Create a business review",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessReviewsController::store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "business_id": {
                                        "description": "id of business",
                                        "type": "string"
                                    },
                                    "score": {
                                        "description": "Score",
                                        "type": "int"
                                    },
                                    "comment": {
                                        "description": "Review comment",
                                        "type": "string"
                                    },
                                    "photo": {
                                        "description": "image",
                                        "type": "file"
                                    },
                                    "mode": {
                                        "description": "Public or private entry.",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "BusinessPostResource"
                    }
                }
            }
        },
...
    }
}